+++ CHANGELOG.md
@@ -5,6 +5,8 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Deploy `index.yaml` to remove composite index for History.
  * NOTE: added daily periodic tasks: `delete-expired-audit-log-records`,
          `delete-expired-consents`, `delete-expired-sessions`.
+ * Removed all usage of `app/static/js/gtag.js` it can be removed after a few
+   runtimeVerions when we are no-longer serving old generated dartdoc files.
 
 ## `20210203t120700-all`